typoe

+++ knowledge4ir/salience/utils.py
@@ -9,7 +9,7 @@ from torch.autograd import Variable
 
 def hinge_loss(output, target):
     assert not target.requires_grad
-    loss = target * (1.0 - output)
+    loss = target.type(torch.FloatTensor) * (1.0 - output)
     loss = loss.clamp(min=0).sum()
     return loss
